Periodic Update
Home page is good, the search page is in pretty good shape, biggest change is the addition of dynamically listing the item categories. Testing has revealed what looks like a problem with the process that links items to categories, the SELECT seems to be ok, need to check the underlying data.